refactor: create manual add defining comments clear app.ts
