+++ styles/components/_modal.scss
@@ -5,8 +5,8 @@
   z-index: 1;
   top: 0;
   left: 0;
-  display: flex;
   // Display & Box Model /
+  display: flex;
   width: 100%;
   height: 100%;
   align-items: center;
@@ -14,9 +14,9 @@
   // Color /
   background-color: rgba(0, 0, 0, 0.8);
   &__container {
+    // Display & Box Model /
     display: flex;
     width: 80%;
-    // Display & Box Model /
     height: 80%;
     flex-direction: column;
     justify-content: space-between;
